regression tests all passing
